# functions/api-specs/v1/openapi.yml
openapi: 3.0.3
info:
  title: Whatboard API
  description: |
    ## Introduction

    The Whatboard API allows you to use your own (or 3rd-party) software to interact with your content on
    Whatboard. You can:

    - Create Boards
    - Update Boards
    - Delete Boards
    - List Boards
    - Retrieve individual Boards

    - Add Blocks to a Board
    - Update Blocks in a Board
    - Delete Blocks from a Board
    - List Blocks in a Board
    - Retrieve individual Blocks in a Board

    ## Limitations

    - It is not currently possible to invite users to collaborate on a Board using this API
    - It is not currently possible to upload files using this API

    ## Authentication

    The Whatboard API uses OpenID Connect to handle authentication.

    The configuration document is located at the following URL: [https://whatboard.app/oidc/.well-known/openid-configuration](https://whatboard.app/oidc/.well-known/openid-configuration)

    You can learn more about OpenID Connect from these sites:

    - [OpenID Connect Playground](https://openidconnect.net/)
    - [OpenID Specification Site](https://openid.net)

    ### Supported flows (`grant_type`s)

    - `implicit`
    - `authorization_code` (with PKCE support)
    - `refresh_token`

    Browser-based clients are encouraged to use the `authorization_code` flow with PKCE. Only use the
    `implicit` flow if using `authorization_code` with PKCE is not possible.

    ### API Docs Note

...
  version: 1.0.0
...
  contact:
...
servers:
...
security:
  - OIDC: []
...
paths:
  /boards:
    get:
      operationId: listBoards
      summary: Retrieve a list of Boards
      description: Retrieve a list of Boards
      tags:
        - boards
      parameters:
        - $ref: "#/components/parameters/OrderBy"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/EqualTo"
        - $ref: "#/components/parameters/Dir"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A list of Boards
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    title: Result
                    type: array
                    items:
                      $ref: "#/components/schemas/BoardOutput"
                  prevPageToken:
                    title: Previous Page Token
                    description: The pagination token for the previous page of results
                    type: string
                  nextPageToken:
                    title: Next Page Token
                    description: The pagination token for the next page of results
                    type: string
        default:
          $ref: "#/components/responses/GenericError"
    post:
      operationId: createBoard
      summary: Create a Board
      description: Create a Board
      tags:
        - boards
      requestBody:
        $ref: "#/components/requestBodies/Board"
      responses:
        "200":
          $ref: "#/components/responses/Board"
        default:
          $ref: "#/components/responses/GenericError"
  "/boards/{boardId}":
    parameters:
      - $ref: "#/components/parameters/BoardId"
    get:
      operationId: getBoard
      summary: Retrieve a single Board by ID
      description: Retrieve a single Board by ID
      tags:
        - boards
      responses:
        "200":
          $ref: "#/components/responses/Board"
        default:
          $ref: "#/components/responses/GenericError"
    patch:
      summary: Update a Board
      requestBody:
        $ref: "#/components/requestBodies/Board"
      responses:
        "200":
          $ref: "#/components/responses/Board"
        default:
          $ref: "#/components/responses/GenericError"
      description: Update a Board
      operationId: updateBoard
      tags:
        - boards
    delete:
      summary: Delete a Board
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/GenericError"
      description: Delete a Board
      operationId: deleteBoard
      tags:
        - boards
  "/blocks/{boardId}":
    parameters:
      - $ref: "#/components/parameters/BoardId"
    get:
      summary: Retrieve a list of Blocks for the given Board
      parameters:
        - $ref: "#/components/parameters/OrderBy"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/EqualTo"
        - $ref: "#/components/parameters/Dir"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A list of Blocks
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    title: Result
                    type: array
                    items:
                      $ref: "#/components/schemas/BlockOutput"
                  prevPageToken:
                    title: Previous Page Token
                    description: The pagination token for the previous page of results
                    type: string
                  nextPageToken:
                    title: Next Page Token
                    description: The pagination token for the next page of results
                    type: string
        default:
          $ref: "#/components/responses/GenericError"
      description: Retrieve a list of Blocks for the given Board
      operationId: listBlocks
      tags:
        - blocks
    post:
      summary: Add a Block to the given Board
      requestBody:
        $ref: "#/components/requestBodies/Block"
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Add a Block to the given Board
      operationId: createBlock
      tags:
        - blocks
  "/blocks/{boardId}/{blockId}":
    parameters:
      - $ref: "#/components/parameters/BoardId"
      - $ref: "#/components/parameters/BlockId"
    get:
      summary: Retrieve a single Block by ID
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Retrieve a single Block by ID
      operationId: getBlock
      tags:
        - blocks
    patch:
      summary: Update a Block
      requestBody:
        $ref: "#/components/requestBodies/Block"
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Update a Block
      operationId: updateBlock
      tags:
        - blocks
    delete:
      summary: Delete a Block
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/GenericError"
      description: Delete a Block
      operationId: deleteBlock
      tags:
        - blocks
...
components:
  schemas:
    BoardInput:
      title: Board (input)
      type: object
      properties:
        board_name:
          title: Board Name
          type: string
          default: my whatboard
        board_header_color:
          title: Header Color
          description: |
            The header color for the Board. The default value is specified in your [Account Branding Settings](/account/branding).
          type: string
          enum:
            - "#dc143c"
            - "#FF2400"
            - "#f67c09"
            - "#ffc106"
            - "#55C57C"
            - "#0da00b"
            - "#0f8df0"
            - "#2c387e"
            - "#051225"
            - "#a043bc"
            - "#F4869C"
            - "#FCEAB4"
            - "#795548"
            - "#4B4B4B"
            - "#1f1f1f"
            - "#000000"
            - "#F1F1F1"
            - "#ffffff00"
            - "#ffffff"
        boardBodyColor:
          title: Body Color
          description: |
            The body color for the Board. The default value is specified in your [Account Branding Settings](/account/branding).
          type: string
          enum:
            - "#dc143c"
            - "#FF2400"
            - "#f67c09"
            - "#ffc106"
            - "#55C57C"
            - "#0da00b"
            - "#0f8df0"
            - "#2c387e"
            - "#051225"
            - "#a043bc"
            - "#F4869C"
            - "#FCEAB4"
            - "#795548"
            - "#4B4B4B"
            - "#1f1f1f"
            - "#000000"
            - "#F1F1F1"
            - "#ffffff00"
            - "#ffffff"
        isPublic:
          title: Publically Accessible
          type: boolean
          default: false
        friendly_url:
          title: Public URL Friendly Code
          type: string
          minLength: 4
          maxLength: 23
        isArchived:
          title: Archived
          description: Whether or not you have archived this Board
          type: boolean
        pinned:
          title: Pinned
          description: Whether or not you have pinned this Board
          type: boolean
    BoardOutput:
      title: Board (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BoardInput"
        - required:
            - id
            - board_members
            - date_created
          properties:
            id:
              title: Board ID
              type: string
            board_members:
              title: Board Members
              type: array
              items:
                type: string
            date_created:
              title: Date Created
              type: string
              format: date-time
            unique_url:
              title: Public URL Invitation Code
              type: string
              format: uuid
            boardUrl:
              title: Board URL
              type: string
              format: url
            boardPublicUrl:
              title: Board Public URL
              description: Only present when `isPublic` is `true`
              type: string
              format: url
            boardPublicFriendlyUrl:
              title: Board Friendly Public URL
              description: Only present when `isPublic` is `true` and `friendly_url` is present
              type: string
              format: url
            isSecured:
              title: Password Protected
              description: |
                Whether or not this Board has been password protected. Passwords are not manageable
                via this API. You must use https://whatboard.app in order to manage Board passwords.
              type: boolean
              default: false
            currentViewingUsers:
              title: Users Currently Viewing Board
              type: object
              additionalProperties:
                x-additionalPropertiesName: User WBID
                type: object
                required:
                  - displayName
                properties:
                  displayName:
                    title: Display Name
                    type: string
                  photoURL:
                    title: Photo URL
                    type: string
                    format: url
            view_logs:
              title: View Logs
              type: object
              additionalProperties:
                x-additionalPropertiesName: User WBID
                type: object
                properties:
                  lastModified:
                    title: Last Modified
                    type: string
                    format: date-time
                  lastViewed:
                    title: Last Viewed
                    type: string
                    format: date-time
            metadata:
              title: Metadata
              type: object
              properties:
                lastModified:
                  title: Last Modified
                  type: string
                  format: date-time
...
    BlockOutput:
      title: Block (output)
      type: object
      oneOf:
        - $ref: "#/components/schemas/TextBlockOutput"
        - $ref: "#/components/schemas/PDFBlockOutput"
        - $ref: "#/components/schemas/ImageBlockOutput"
        - $ref: "#/components/schemas/VideoBlockOutput"
        - $ref: "#/components/schemas/ButtonsBlockOutput"
        - $ref: "#/components/schemas/ChecklistBlockOutput"
        - $ref: "#/components/schemas/ConversationBlockOutput"
        - $ref: "#/components/schemas/ScriptEmbedBlockOutput"
        - $ref: "#/components/schemas/IframeBlockOutput"
        - $ref: "#/components/schemas/FilesBlockOutput"
        - $ref: "#/components/schemas/QAFormBlockOutput"
        - $ref: "#/components/schemas/RSSBlockOutput"
        - $ref: "#/components/schemas/FileRequestBlockOutput"
        - $ref: "#/components/schemas/GridBlockOutput"
      discriminator:
        propertyName: type
        x-explicitMappingOnly: true
        mapping:
          Text: "#/components/schemas/TextBlockOutput"
          PDF: "#/components/schemas/PDFBlockOutput"
          Image: "#/components/schemas/ImageBlockOutput"
          Video: "#/components/schemas/VideoBlockOutput"
          Buttons: "#/components/schemas/ButtonsBlockOutput"
          Checklist: "#/components/schemas/ChecklistBlockOutput"
          Conversation: "#/components/schemas/ConversationBlockOutput"
          ScriptEmbed: "#/components/schemas/ScriptEmbedBlockOutput"
          Iframe: "#/components/schemas/IframeBlockOutput"
          Files: "#/components/schemas/FilesBlockOutput"
          QAForm: "#/components/schemas/QAFormBlockOutput"
          RSS: "#/components/schemas/RSSBlockOutput"
          FileRequest: "#/components/schemas/FileRequestBlockOutput"
          Grid: "#/components/schemas/GridBlockOutput"
...
    TextBlockOutput:
      title: Text (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/TextBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    PDFBlockOutput:
      title: PDF (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/PDFBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    ImageBlockOutput:
      title: Image (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ImageBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    VideoBlockOutput:
      title: Video (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/VideoBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    ButtonsBlockOutput:
      title: Buttons (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ButtonsBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    ChecklistBlockOutput:
      title: Checklist (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ChecklistBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    ConversationBlockOutput:
      title: Chat (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ConversationBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    ScriptEmbedBlockOutput:
      title: Embed HTML (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ScriptEmbedBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    IframeBlockOutput:
      title: Embed Web Page (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/IframeBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    FilesBlockOutput:
      title: File Storage (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/FilesBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    QAFormBlockOutput:
      title: Q & A Form (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/QAFormBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    RSSBlockOutput:
      title: RSS Feed (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/RSSBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    Error:
      title: Error
      type: object
      required:
        - error
      properties:
        error:
          title: Error Message
          type: string
        errorDescription:
          title: Error Description
          type: string
  parameters:
    BoardId:
      name: boardId
      in: path
      schema:
        type: string
      required: true
    BlockId:
      name: blockId
      in: path
      schema:
        type: string
      required: true
    OrderBy:
      name: orderBy
      in: query
      schema:
        type: string
    PageToken:
      name: pageToken
      in: query
      schema:
        type: string
    EqualTo:
      name: equalTo
      in: query
      schema:
        type: string
    Dir:
      name: dir
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
  requestBodies:
    Board:
      description: A single Board
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BoardInput"
    Block:
      description: A single Block
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BlockInput"
  responses:
    Board:
      description: A single Board
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BoardOutput"
    Block:
      description: A single Block
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BlockOutput"
...
    NoContent:
      description: No content
    GenericError:
      description: Generic Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    OIDC:
      type: openIdConnect
      openIdConnectUrl: "https://whatboard.app/oidc/.well-known/openid-configuration"